[rename] external -> common으로 변경 (#46)
