+++ src/main/java/org/recordy/server/common/exception/ExternalException.java
@@ -1,4 +1,4 @@
-package org.recordy.server.external.exception;
+package org.recordy.server.common.exception;
 
 import org.recordy.server.common.exception.RecordyException;
 import org.recordy.server.common.message.ErrorMessage;
