+++ src/main/java/org/recordy/server/common/service/S3Service.java
@@ -1,4 +1,4 @@
-package org.recordy.server.external.service;
+package org.recordy.server.common.service;
 
 import org.springframework.web.multipart.MultipartFile;
