+++ src/test/java/org/recordy/server/external/controller/S3TestControllerTest.java
@@ -2,7 +2,8 @@ package org.recordy.server.external.controller;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.recordy.server.external.service.impl.S3ServiceImpl;
+import org.recordy.server.common.controller.S3TestController;
+import org.recordy.server.common.service.impl.S3ServiceImpl;
 import org.recordy.server.mock.FakeContainer;
 import org.springframework.mock.web.MockMultipartFile;
