+++ src/test/java/org/recordy/server/external/service/S3ServiceTest.java
@@ -2,11 +2,9 @@ package org.recordy.server.external.service;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.recordy.server.external.config.S3Config;
-import org.recordy.server.external.exception.ExternalException;
-import org.recordy.server.external.service.impl.S3ServiceImpl;
-import org.recordy.server.mock.FakeContainer;
-import org.recordy.server.user.exception.UserException;
+import org.recordy.server.common.config.S3Config;
+import org.recordy.server.common.exception.ExternalException;
+import org.recordy.server.common.service.impl.S3ServiceImpl;
 import org.springframework.web.multipart.MultipartFile;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
